catatan/requirements.txt
google-generativeai
streamlit
pandas
plotly
python-dotenv
requests